Improve testing; handle empty upstream git repos.
Reworked the testing so there's actually a bare/upstream repo. While
I was at it, I had to add handling the case of a bare git repo with
no commits; in this event there's no branches yet. So, now we check
for that and create an initial commit with a README.md.