Unbreak symlinking, fix edge case in branch name for bare repos.
Symlinking requires a little more than just an optional arg, i.e. we
need to make sure we're copying the original file into the repo first
before creating the symlink in the home directory.

Also, found an edge case where some systems will still default to
'master' for the git branch, which breaks the initial commit/push when
dealing with a brand new bare origin. So, we rename the local branch to
'main' in this case.