Fix edge-case with realpath
If you use realpath on a nonexistent file, it's OK as long
as the directories exist. But if you try to realpath a
file in a directory that doesn't exist, realpath fails.

This behavior was masking a problem when trying to add
or status-check a new file that didn't already have a
directory in the repo.